// src/interfaces/appInterfaces.tsx
// Generated by https://quicktype.io
export interface LoginData{
  correo:string;
  password:string;
}

export interface LoginResponse {
  usuario:Usuario;
  token:string;
}

export interface RegisterData{
  nombre:string;
  correo:string;
  password:string;
  rol:'USER_ROLE'|'ADMIN_ROLE'|'VENTAS_ROLE';
}

export interface RegisterResponse {
  usuario:Usuario;
  token:string;
}

export interface Usuario {
  rol:string;
  estado:boolean;
  google:boolean;
  nombre:string;
  correo:string;
  role:string;
  uid:string;
  img?:string;
}


//Productos 

// Generated by https://quicktype.io

export interface ProductsResponse {
  total:number;
  productos:Producto[];
}

export interface Producto {
  precio:number;
  _id:string;
  nombre:string;
  usuario:Categoria;
  categoria:Categoria;
  disponible:boolean;
}

export interface Categoria {
  _id:string;
  nombre:string;
}

//ProductContext Types
export interface IAddProduct{
  categoryId:string;
  productNane:string
}

export interface IUpdateProduct{
  categoryId:string;
  productName:string;
  productId:string;
}

export interface IDeleteProduct{
  id:string
}

export interface ILoadProductById{
  id:string
}

export interface IUploadImage{
  data:any;
  id:string;
}